requirements.txt
pandas
yfinance
curl_cffi
beautifulsoup4
lxml
aiohttp
aiofiles
mplfinance
textblob
streamlit
plotly
sqlalchemy
openpyxl
xlsxwriter
schedule
python-dotenv
